+++ docs/background/make_reflected_regions.py
@@ -1,22 +1,21 @@
+"""Example how to compute and plot reflected regions."""
 from astropy.coordinates import SkyCoord, Angle
-from astropy.wcs import WCS
-
-from gammapy.region import find_reflected_regions
-from gammapy.extern.regions.shapes import CircleSkyRegion
+from gammapy.extern.regions import CircleSkyRegion
 from gammapy.image import ExclusionMask
-from gammapy.utils.testing import requires_data
+from gammapy.background import find_reflected_regions
 
 mask = ExclusionMask.empty(name='Exclusion Mask', nxpix=801, nypix=701, binsz=0.01,
-                       	   coordsys='CEL', xref=83.2, yref=22.7)
+                           coordsys='CEL', xref=83.2, yref=22.7)
 mask.fill_random_circles(n=8, min_rad=30, max_rad=80)
 
-pos = SkyCoord(80.2, 23.5, unit='deg', frame='icrs')
+pos = SkyCoord(80.2, 23.5, unit='deg')
 radius = Angle(0.4, 'deg')
 test_region = CircleSkyRegion(pos, radius)
-center = SkyCoord(82.8, 22.5, unit='deg', frame='icrs')
+center = SkyCoord(82.8, 22.5, unit='deg')
 regions = find_reflected_regions(test_region, center, mask)
 
 import matplotlib.pyplot as plt
+
 fig = plt.figure(figsize=(8, 5), dpi=80)
 ax = fig.add_axes([0.1, 0.1, 0.8, 0.8], projection=mask.wcs)
 mask.plot(ax, fig)