+++ gammapy/background/__init__.py
@@ -12,3 +12,4 @@ from .template import *
 from .kernel import *
 from .models import *
 from .off_data_background_maker import *
+from .reflected import *