+++ gammapy/background/reflected.py
@@ -1,8 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from __future__ import absolute_import, division, print_function, unicode_literals
 import numpy as np
-from . import PixRegionList, PixCircleRegion
 from astropy.coordinates import Angle
+from ..extern.regions import PixCoord, CirclePixelRegion
 
 __all__ = [
     'find_reflected_regions',
@@ -17,12 +17,12 @@ def find_reflected_regions(region, center, exclusion_mask, angle_increment=None,
 
     Parameters
     ----------
-    region : `~gammapy.region.Region`
+    region : `~regions.CircleSkyRegion`
         Region
     center : `~astropy.coordinates.SkyCoord`
         Rotation point
     exclusion_mask : `~gammapy.image.ExclusionMask`
-        Exlusion mask
+        Exclusion mask
     angle_increment : `~astropy.coordinates.Angle`
         Rotation angle for each step
     min_distance : `~astropy.coordinates.Angle`
@@ -32,36 +32,35 @@ def find_reflected_regions(region, center, exclusion_mask, angle_increment=None,
 
     Returns
     -------
-    regions : `~gammapy.region.SkyRegionList`
+    regions : list of `~regions.SkyRegion`
         Reflected regions list
     """
     angle_increment = Angle('0.1 rad') if angle_increment is None else Angle(angle_increment)
     min_distance = Angle('0 rad') if min_distance is None else Angle(min_distance)
     min_distance_input = Angle('0 rad') if min_distance_input is None else Angle(min_distance_input)
 
-    reflected_regions_pix = PixRegionList()
+    reflected_regions_pix = list()
     wcs = exclusion_mask.wcs
     pix_region = region.to_pixel(wcs)
-    val = center.to_pixel(wcs, origin=1)
-    pix_center = (float(val[0]), float(val[1]))
-    dx = pix_region.center[0] - pix_center[0]
-    dy = pix_region.center[1] - pix_center[1]
+    pix_center = PixCoord(*center.to_pixel(wcs, origin=1))
+    dx = pix_region.center[0] - pix_center.x
+    dy = pix_region.center[1] - pix_center.y
     offset = np.hypot(dx, dy)
     angle = Angle(np.arctan2(dx, dy), 'rad')
-    min_ang = Angle(2 * pix_region.radius / offset, 'rad') + min_distance
+    min_ang = Angle(2 * pix_region.radius.to('pix').value / offset, 'rad') + min_distance
     max_angle = angle + Angle('360deg') - min_ang - min_distance_input
 
     curr_angle = angle + min_ang + min_distance_input
     while curr_angle < max_angle:
         test_pos = _compute_xy(pix_center, offset, curr_angle)
-        test_reg = PixCircleRegion(test_pos, pix_region.radius)
-        if not test_reg.is_inside_exclusion(exclusion_mask):
+        test_reg = CirclePixelRegion(test_pos, pix_region.radius)
+        if not _is_inside_exclusion(test_reg, exclusion_mask):
             reflected_regions_pix.append(test_reg)
             curr_angle = curr_angle + min_ang
         else:
             curr_angle = curr_angle + angle_increment
 
-    reflected_regions = reflected_regions_pix.to_sky(wcs)
+    reflected_regions = [_.to_sky(wcs) for _ in reflected_regions_pix]
     return reflected_regions
 
 
@@ -72,6 +71,14 @@ def _compute_xy(pix_center, offset, angle):
     """
     dx = offset * np.sin(angle)
     dy = offset * np.cos(angle)
-    x = pix_center[0] + dx
-    y = pix_center[1] + dy
+    x = pix_center.x + dx
+    y = pix_center.y + dy
     return x, y
+
+# TODO :Copied from gammapy.region.PixCircleList (deleted), find better place
+def _is_inside_exclusion(pixreg, exclusion):
+    x, y = pixreg.center
+    excl_dist = exclusion.distance_image
+    val = excl_dist[np.round(y).astype(int), np.round(x).astype(int)]
+    return val < pixreg.radius.to('pix').value
+