+++ gammapy/background/tests/test_models.py
@@ -9,13 +9,13 @@ import astropy.units as u
 from astropy.units import Quantity
 from astropy.coordinates import Angle, SkyCoord
 from astropy.modeling.models import Gaussian1D
+from ...extern.regions import CircleSkyRegion
 from ...utils.testing import requires_dependency, requires_data
 from ...datasets import gammapy_extra
 from ...background import GaussianBand2D, CubeBackgroundModel, EnergyOffsetBackgroundModel
 from ...utils.energy import EnergyBounds
 from ...data import ObservationTable
 from ...data import DataStore, EventList
-from ...region import SkyCircleRegion
 from ...background.models import _compute_pie_fraction, _select_events_outside_pie
 
 
@@ -156,10 +156,10 @@ def make_test_array_oneobs(excluded_sources=None, fov_radius=Angle(2.5, "deg")):
 def make_excluded_sources():
     centers = SkyCoord([1, 0], [2, 1], unit='deg')
     radius = Angle('0.3 deg')
-    sources = SkyCircleRegion(pos=centers, radius=radius)
+    sources = CircleSkyRegion(centers, radius)
     catalog = Table()
-    catalog["RA"] = sources.pos.data.lon
-    catalog["DEC"] = sources.pos.data.lat
+    catalog["RA"] = sources.center.data.lon
+    catalog["DEC"] = sources.center.data.lat
     catalog["Radius"] = sources.radius
     return catalog
 
@@ -167,10 +167,10 @@ def make_excluded_sources():
 def make_source_nextCrab():
     center = SkyCoord([84, 89], [23, 20], unit='deg', frame='icrs')
     radius = Angle('0.3 deg')
-    sources = SkyCircleRegion(pos=center, radius=radius)
+    sources = CircleSkyRegion(center, radius)
     catalog = Table()
-    catalog["RA"] = sources.pos.data.lon
-    catalog["DEC"] = sources.pos.data.lat
+    catalog["RA"] = sources.center.data.lon
+    catalog["DEC"] = sources.center.data.lat
     catalog["Radius"] = sources.radius
     return catalog