+++ gammapy/data/tests/test_observation_stats.py
@@ -4,17 +4,16 @@ from numpy.testing import assert_allclose
 from astropy.tests.helper import pytest
 from astropy.coordinates import SkyCoord
 import astropy.units as u
+from ...extern.regions import CircleSkyRegion
 from ...data import DataStore, ObservationList, ObservationStats, Target
 from ...utils.testing import requires_data, requires_dependency
-from ...extern.regions.shapes import CircleSkyRegion
 from ...background import reflected_regions_background_estimate as refl
 from ...image import ExclusionMask
 
 
 @requires_data('gammapy-extra')
 def get_obs_list():
-    data_store = DataStore.from_dir(
-        '$GAMMAPY_EXTRA/datasets/hess-crab4-hd-hap-prod2/')
+    data_store = DataStore.from_dir('$GAMMAPY_EXTRA/datasets/hess-crab4-hd-hap-prod2/')
     run_list = [23523, 23526]
     obs_list = ObservationList([data_store.obs(_) for _ in run_list])
     return obs_list
@@ -30,7 +29,7 @@ def get_obs(id):
 
 @pytest.fixture
 def target():
-    pos = SkyCoord(83.63 * u.deg, 22.01 * u.deg, frame='icrs')
+    pos = SkyCoord(83.63 * u.deg, 22.01 * u.deg)
     on_size = 0.3 * u.deg
     on_region = CircleSkyRegion(pos, on_size)
 
@@ -44,8 +43,7 @@ def target():
 @requires_data('gammapy-extra')
 @pytest.fixture
 def get_mask():
-    mask = ExclusionMask.read(
-        '$GAMMAPY_EXTRA/datasets/exclusion_masks/tevcat_exclusion.fits')
+    mask = ExclusionMask.read('$GAMMAPY_EXTRA/datasets/exclusion_masks/tevcat_exclusion.fits')
     return mask
 
 
@@ -70,5 +68,5 @@ def test_stack(target):
                   run.pointing_radec, get_mask(), run.events)
         obs_stats.append(ObservationStats.from_target(run, target, bg))
     sum_obs_stats = ObservationStats.stack(obs_stats)
-    assert_allclose(sum_obs_stats.alpha, 0.284, rtol=1.e-2)
-    assert_allclose(sum_obs_stats.sigma, 23.48, rtol=1.e-3)
+    assert_allclose(sum_obs_stats.alpha, 0.284, rtol=1e-2)
+    assert_allclose(sum_obs_stats.sigma, 23.35, rtol=1e-3)