+++ gammapy/extern/regions/core/__init__.py
@@ -1,2 +1,6 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Region core functionality.
+"""
 from .core import *
 from .pixcoord import *
+from .compound import *