+++ gammapy/extern/regions/core/compound.py
@@ -1,4 +1,8 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+from __future__ import absolute_import, division, print_function, unicode_literals
 from .core import PixelRegion, SkyRegion
+
+__all__ = ['CompoundPixelRegion', 'CompoundSkyRegion']
 
 
 class CompoundPixelRegion(PixelRegion):