+++ gammapy/extern/regions/core/pixcoord.py
@@ -1,7 +1,14 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+__all__ = ['PixCoord']
+
+
 class PixCoord(object):
     """
     Class representing a collection of pixel coordinates
     """
+
     def __init__(self, x, y):
         self.x = x
         self.y = y