+++ gammapy/extern/regions/io/__init__.py
@@ -1,5 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""I/O
+"""Region I/O.
 """
 from .read_ds9 import *
 from .write_ds9 import *