+++ gammapy/extern/regions/shapes/__init__.py
@@ -1,7 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""Region shapes
+"""Region shapes.
 """
 from .circle import *
 from .ellipse import *
+from .point import *
 from .polygon import *
 from .rectangle import *