+++ gammapy/extern/regions/shapes/circle.py
@@ -1,15 +1,12 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-
+from __future__ import absolute_import, division, print_function, unicode_literals
 import math
-
 import numpy as np
-
-from astropy import wcs
-from astropy import coordinates
-from astropy import units as u
-
+from astropy.wcs.utils import pixel_to_skycoord
 from ..core import PixelRegion, SkyRegion
 from ..utils.wcs_helpers import skycoord_to_pixel_scale_angle
+
+__all__ = ['CirclePixelRegion', 'CircleSkyRegion']
 
 
 class CirclePixelRegion(PixelRegion):
@@ -42,13 +39,21 @@ class CirclePixelRegion(PixelRegion):
     def to_shapely(self):
         return self.center.to_shapely().buffer(self.radius)
 
-    def to_sky(self, mywcs, mode='local', tolerance=None):
-        # TODO: needs to be implemented
-        raise NotImplementedError("")
+    def to_sky(self, wcs, mode='local', tolerance=None):
+        if mode != 'local':
+            raise NotImplementedError
+        if tolerance is not None:
+            raise NotImplementedError
+
+        skypos = pixel_to_skycoord(self.center[0], self.center[1], wcs)
+        xc, yc, scale, angle = skycoord_to_pixel_scale_angle(skypos, wcs)
+
+        radius_sky = self.radius / scale
+        return CircleSkyRegion(skypos, radius_sky)
 
     def to_mask(self, mode='center'):
         # TODO: needs to be implemented
-        raise NotImplementedError("")
+        raise NotImplementedError
 
     def as_patch(self, **kwargs):
         import matplotlib.patches as mpatches
@@ -89,14 +94,14 @@ class CircleSkyRegion(SkyRegion):
         rad = self.radius
         return '{clsnm}\nCenter:{coord}\nRadius:{rad}'.format(**locals())
 
-    def to_pixel(self, mywcs, mode='local', tolerance=None):
+    def to_pixel(self, wcs, mode='local', tolerance=None):
         """
         Given a WCS, convert the circle to a best-approximation circle in pixel
         dimensions.
 
         Parameters
         ----------
-        mywcs : `~astropy.wcs.WCS`
+        wcs : `~astropy.wcs.WCS`
             A world coordinate system
         mode : 'local' or not
             not implemented
@@ -109,12 +114,12 @@ class CircleSkyRegion(SkyRegion):
         """
 
         if mode != 'local':
-            raise NotImplementedError()
+            raise NotImplementedError
         if tolerance is not None:
-            raise NotImplementedError()
+            raise NotImplementedError
 
-        xc, yc, scale, angle = skycoord_to_pixel_scale_angle(self.center, mywcs)
-        radius_pix = (self.radius * scale).to(u.pixel).value
+        xc, yc, scale, angle = skycoord_to_pixel_scale_angle(self.center, wcs)
+        radius_pix = (self.radius * scale)
 
         pixel_positions = np.array([xc, yc]).transpose()