+++ CHANGES.rst
@@ -22,6 +22,12 @@ Gammapy 0.5 will be released on July 7, 2016.
 
 For plans and progress see https://github.com/gammapy/gammapy/milestones/0.5
 
+Highlights:
+
+- Regions: http://astropy-regions.readthedocs.io/
+- Image and cube analysis improvements
+- Time analysis
+
 Contributors
 ++++++++++++
 
@@ -44,6 +50,7 @@ This list is incomplete. Many small improvements and bug fixes are not listed he
 
 See the complete `Gammapy 0.5 merged pull requests list on Github <https://github.com/gammapy/gammapy/pulls?utf8=%E2%9C%93&q=is%3Apr+milestone%3A0.5+is%3Amerged+>`__.
 
+- [#558] Finish change to use gammapy.extern.regions (Johannes King and Christoph Deil)
 - [#569] Add detection utilities à la BgStats (Julien Lefaucheur)
 - [#565] Add exptest time variability test (Helen Poon)
 - [#564] Add LightCurve class (Nachiketa Chakraborty)